Update: Android 14 Support

+++ apptentive-feedback/src/main/java/apptentive/com/android/feedback/utils/FileUtil.kt
@@ -340,11 +340,7 @@ object FileUtil {
     fun deleteFile(filePath: String?) {
         if (!filePath.isNullOrBlank()) {
             val file = File(filePath)
-            if (file.exists())
-                file.delete()
+            if (file.exists()) file.delete()
         }
     }
-
-    fun isConversationCacheStoredInLegacyFormat(filePath: String): Boolean =
-        filePath.contains("apptentive/conversations")
 }
